Invoke Test mode at run time with a URL argument.
There is an option to set the Payment Processor into Test mode when configuring the Contribution section of the Webform. 

This change adds the ability also to set Test mode at run time, with a URL argument of '?payment=test'. This enables one to run regression tests for alterations made to the live site without having to take it out of service to change the configuration parameter, with the inconvenience and risk of error involved in doing so.